requirements.txt
pytube==15.0.0
--find-links https://download.pytorch.org/whl/torch_stable.html
numpy==1.24.3
torch==2.2.0+cpu
torchaudio==2.2.0+cpu
torchvision==0.17.0+cpu
openai-whisper
ffmpeg-python==0.2.0
tqdm==4.66.1
youtube-transcript-api
python-dotenv
requests